AI/Agentic AI/2. PhiData Playground Application/requirements.txt
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
psutil